Move some functions to misc.py
