+++ bitcoinx/misc.py
@@ -9,14 +9,19 @@
 
 __all__ = (
     'be_bytes_to_int', 'le_bytes_to_int',
-    'int_to_be_bytes', 'int_to_le_bytes', 'CONTEXT'
+    'int_to_be_bytes', 'int_to_le_bytes', 'CONTEXT',
+    'is_valid_hostname', 'classify_host', 'validate_port', 'validate_protocol',
+    'NetAddress', 'Service', 'ServicePart',
 )
 
-import mmap
+import re
+from enum import IntEnum
 from functools import partial
+from ipaddress import ip_address, IPv4Address, IPv6Address
 from os import path
-
 from electrumsv_secp256k1 import create_context
+
+from .packing import pack_port
 
 
 CONTEXT = create_context()
@@ -52,21 +57,241 @@ def data_file_path(*parts):
     return path.join(package_dir, "data", *parts)
 
 
+class ServicePart(IntEnum):
+    PROTOCOL = 0
+    HOST = 1
+    PORT = 2
+
+
+class NetAddress:
+
+    def __init__(self, host, port, check_port=True):
+        '''Construct a NetAddress from a host and a port.
+
+        Host is classified and port is an integer.'''
+        self._host = classify_host(host)
+        self._port = validate_port(port) if check_port else int(port)
+
+    def __eq__(self, other):
+        # pylint: disable=protected-access
+        return (isinstance(other, NetAddress) and
+                self._host == other._host and self._port == other._port)
+
+    def __hash__(self):
+        return hash((self._host, self._port))
+
+    @classmethod
+    def from_string(cls, string, *, check_port=True, default_func=None):
+        '''Construct a NetAddress from a string and return a (host, port) pair.
+
+        If either (or both) is missing and default_func is provided, it is called with
+        ServicePart.HOST or ServicePart.PORT to get a default.
+        '''
+        def split_address(string):
+            if string.startswith('['):
+                end = string.find(']')
+                if end != -1:
+                    if len(string) == end + 1:
+                        return string[1:end], ''
+                    if string[end + 1] == ':':
+                        return string[1:end], string[end+2:]
+            colon = string.find(':')
+            if colon == -1:
+                return string, ''
+            return string[:colon], string[colon + 1:]
+
+        if not isinstance(string, str):
+            raise TypeError(f'address must be a string: {string}')
+        host, port = split_address(string)
+        if default_func:
+            host = host or default_func(ServicePart.HOST)
+            port = port or default_func(ServicePart.PORT)
+            if not host or not port:
+                raise ValueError(f'invalid address string: {string}')
+        return cls(host, port, check_port=check_port)
+
+    @property
+    def host(self):
+        return self._host
+
+    @property
+    def port(self):
+        return self._port
+
+    def __str__(self):
+        if isinstance(self._host, IPv6Address):
+            return f'[{self._host}]:{self._port}'
+        return f'{self.host}:{self.port}'
+
+    def __repr__(self):
+        return f"NetAddress('{self}')"
+
+    @classmethod
+    def default_host_and_port(cls, host, port):
+        def func(kind):
+            return host if kind == ServicePart.HOST else port
+        return func
+
+    @classmethod
+    def default_host(cls, host):
+        return cls.default_host_and_port(host, None)
+
+    @classmethod
+    def default_port(cls, port):
+        return cls.default_host_and_port(None, port)
+
+    def pack_host(self):
+        '''Return the host as a 16-byte IPv6 address.'''
+        if isinstance(self._host, IPv4Address):
+            # An IPv4-mapped IPv6 address
+            return b'\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\xff\xff' + self._host.packed
+        if isinstance(self._host, IPv6Address):
+            return self._host.packed
+        raise TypeError(f'address must be resolved: {self._host}')
+
+    def pack(self):
+        return self.pack_host() + pack_port(self._port)
+
+
+class Service:
+    '''A validated protocol, address pair.'''
+
+    def __init__(self, protocol, address):
+        '''Construct a service from a protocol string and a NetAddress object,'''
+        self._protocol = validate_protocol(protocol)
+        if not isinstance(address, NetAddress):
+            address = NetAddress.from_string(address)
+        self._address = address
+
+    def __eq__(self, other):
+        # pylint: disable=protected-access
+        return (isinstance(other, Service) and
+                self._protocol == other._protocol and self._address == other._address)
+
+    def __hash__(self):
+        return hash((self._protocol, self._address))
+
+    @property
+    def protocol(self):
+        return self._protocol
+
+    @property
+    def address(self):
+        return self._address
+
+    @property
+    def host(self):
+        return self._address.host
+
+    @property
+    def port(self):
+        return self._address.port
+
+    @classmethod
+    def from_string(cls, string, *, default_func=None):
+        '''Construct a Service from a string.
+
+        If default_func is provided and any ServicePart is missing, it is called with
+        default_func(protocol, part) to obtain the missing part.
+        '''
+        if not isinstance(string, str):
+            raise TypeError(f'service must be a string: {string}')
+
+        parts = string.split('://', 1)
+        if len(parts) == 2:
+            protocol, address = parts
+        else:
+            item, = parts
+            protocol = None
+            if default_func:
+                if default_func(item, ServicePart.HOST) and default_func(item, ServicePart.PORT):
+                    protocol, address = item, ''
+                else:
+                    protocol, address = default_func(None, ServicePart.PROTOCOL), item
+            if not protocol:
+                raise ValueError(f'invalid service string: {string}')
+
+        if default_func:
+            default_func = partial(default_func, protocol.lower())
+        address = NetAddress.from_string(address, default_func=default_func)
+        return cls(protocol, address)
+
+    def __str__(self):
+        return f'{self._protocol}://{self._address}'
+
+    def __repr__(self):
+        return f"Service({self._protocol!r}, '{self._address}')"
+
+
+#
+# Network utilities
+#
+
+
+# See http://stackoverflow.com/questions/2532053/validate-a-hostname-string
+# Note underscores are valid in domain names, but strictly invalid in host
+# names.  We ignore that distinction.
+PROTOCOL_REGEX = re.compile('[A-Za-z][A-Za-z0-9+-.]+$')
+LABEL_REGEX = re.compile('^[a-z0-9_]([a-z0-9-_]{0,61}[a-z0-9_])?$', re.IGNORECASE)
+NUMERIC_REGEX = re.compile('[0-9]+$')
+
+
+def is_valid_hostname(hostname):
+    '''Return True if hostname is valid, otherwise False.'''
+    if not isinstance(hostname, str):
+        raise TypeError('hostname must be a string')
+    # strip exactly one dot from the right, if present
+    if hostname and hostname[-1] == ".":
+        hostname = hostname[:-1]
+    if not hostname or len(hostname) > 253:
+        return False
+    labels = hostname.split('.')
+    # the TLD must be not all-numeric
+    if re.match(NUMERIC_REGEX, labels[-1]):
+        return False
+    return all(LABEL_REGEX.match(label) for label in labels)
+
+
+def classify_host(host):
+    '''Host is an IPv4Address, IPv6Address or a string.
+
+    If an IPv4Address or IPv6Address return it.  Otherwise convert the string to an
+    IPv4Address or IPv6Address object if possible and return it.  Otherwise return the
+    original string if it is a valid hostname.
+
+    Raise ValueError if a string cannot be interpreted as an IP address and it is not
+    a valid hostname.
+    '''
+    if isinstance(host, (IPv4Address, IPv6Address)):
+        return host
+    if is_valid_hostname(host):
+        return host
+    return ip_address(host)
+
+
+def validate_port(port):
+    '''Validate port and return it as an integer.
+
+    A string, or its representation as an integer, is accepted.'''
+    if not isinstance(port, (str, int)):
+        raise TypeError(f'port must be an integer or string: {port}')
+    if isinstance(port, str) and port.isdigit():
+        port = int(port)
+    if isinstance(port, int) and 0 < port <= 65535:
+        return port
+    raise ValueError(f'invalid port: {port}')
+
+
+def validate_protocol(protocol):
+    '''Validate a protocol, a string, and return it in lower case.'''
+    if not re.match(PROTOCOL_REGEX, protocol):
+        raise ValueError(f'invalid protocol: {protocol}')
+    return protocol.lower()
+
+
 #
 # Internal utilities
 #
-
-def map_file(file_name, new_size=None):
-    '''Map an existing file into memory.  If new_size is specified the
-    file is truncated or extended to that size.
-
-    Returns a Python mmap object.
-    '''
-    with open(file_name, 'rb+') as f:
-        if new_size is not None:
-            f.truncate(new_size)
-        return mmap.mmap(f.fileno(), 0)
-
 
 # Method decorator.  To be used for calculations that will always deliver the same result.
 # The method cannot take any arguments and should be accessed as an attribute.
